Initial item select and item leave
The player may now look for and interact with items around the scene. Items that are hovered over now highlight and display their name in the bottom right corner of the HUD screen. The player may select the object, which brings them to a menu in which they can read it's description. In this menu, they may also rotate the object by clicking and dragging anywhere on screen (besides the menu buttons). The player may leave the object, which will close the description screen and return the player to the game world.